refactor: colocate realtime tests
